section files moved to pages/home
