+++ src/pages/Home/Index.tsx
@@ -1,14 +1,14 @@
 import { BannerCarousel } from "@/components/carousels/BannerCarousel";
 import BookingForm from "@/components/forms/BookingForm";
-import ImageCarousel from '../components/imagecarousel/ImageCarousel';
-import About from "../components/sections/About";
-import ClientReviews  from "../components/sections/ClientReviews";
+import ImageCarousel from '../../components/imagecarousel/ImageCarousel';
+import About from "./About";
+import ClientReviews  from "./ClientReviews";
 // import banner from '../assets/img/marcelinos-banner.jpg';
-import EventVenues from "../components/cards/EventVenue";
+import EventVenues from "../../components/cards/EventVenue";
 import RoomCard from "@/components/cards/RoomCard";
-import FAQ from "../components/sections/ContactForm";
+import FAQ from "./ContactForm";
 import LocationMap from "@/components/map/LocationMap";
-import Services from "../components/cards/Services";
+import Services from "../../components/cards/Services";
