+++ src/routes/route.ts
@@ -1,7 +1,7 @@
 // src/routes/route.js
 import { lazy } from "react";
 
-const Home = lazy(() => import("../pages/Home"));
+const Home = lazy(() => import("../pages/Home/Index"));
 const Rooms = lazy(() => import("../pages/Rooms"));
 const NotFound = lazy(() => import("../pages/NotFound"));
 const PrivacyPolicy = lazy(() => import("../pages/PrivacyPolicy"));
